requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
numpy==1.24.3
pandas==1.5.3
pytz==2023.3
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
